requirements.txt
Flask
python-dotenv
requests
flask_sqlalchemy
SQLAlchemy
psycopg2
json
